move atomic_wrappers to examples directory
