+++ creusot-std/src/std/sync.rs
@@ -1,5 +1,4 @@
 pub mod atomic;
-pub mod atomic_no_LA;
 #[cfg(feature = "sc-drf")]
 pub mod atomic_sc;
 pub mod committer;
