+++ examples/atomic_wrappers/atomic_wrappers.stderr
@@ -1,0 +1,21 @@
+warning: unused imports: `AcqRel`, `Acquire`, `AtomicI16`, `AtomicI32`, `AtomicI64`, `AtomicI8`, `AtomicIsize`, `AtomicPtr`, `AtomicU16`, `AtomicU32`, `AtomicU64`, `AtomicU8`, `AtomicUsize`, `FnGhost`, `Relaxed`, and `Release`
+  --> atomic_wrappers.rs:6:13
+   |
+ 6 | ...::{FnGhost, Perm, perm::PermTarget},
+   |       ^^^^^^^
+...
+12 | ...tomic::{ordering::{Ordering, LoadOrdering, StoreOrdering, UpdateOrdering, Relaxed, Acquire, Release, AcqRel}, AtomicBool, AtomicPtr, AtomicI8, AtomicU8, AtomicI16, AtomicI32, AtomicI64, AtomicIsize, AtomicU16, AtomicU32, AtomicU64, AtomicUsize},
+   |                                                                              ^^^^^^^  ^^^^^^^  ^^^^^^^  ^^^^^^               ^^^^^^^^^  ^^^^^^^^  ^^^^^^^^  ^^^^^^^^^  ^^^^^^^^^  ^^^^^^^^^  ^^^^^^^^^^^  ^^^^^^^^^  ^^^^^^^^^  ^^^^^^^^^  ^^^^^^^^^^^
+   |
+   = note: `#[warn(unused_imports)]` (part of `#[warn(unused)]`) on by default
+
+warning: trait `AtomicExt` is never used
+  --> atomic_wrappers.rs:17:7
+   |
+17 | trait AtomicExt<'a, BaseType : DeepModel> : Sized + PermTarget<Value<'a> = FMap<Timestamp, (...
+   |       ^^^^^^^^^
+   |
+   = note: `#[warn(dead_code)]` (part of `#[warn(unused)]`) on by default
+
+warning: 2 warnings emitted
+
