+++ examples/atomic_wrappers/atomic_wrappers/why3session.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE why3session PUBLIC "-//Why3//proof session v5//EN"
+"https://www.why3.org/why3session.dtd">
+<why3session shape_version="6">
+<prover id="0" name="Z3" version="" timelimit="0.2" steplimit="0" memlimit="1000"/>
+<file format="coma">
+<path name=".."/><path name="atomic_wrappers.coma"/>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_compare_exchange">
+ <goal name="vc_wrap_compare_exchange_AtomicBool">
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_compare_exchange__refines">
+ <goal name="refines">
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_compare_exchange_weak">
+ <goal name="vc_wrap_compare_exchange_weak_AtomicBool">
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_compare_exchange_weak__refines">
+ <goal name="refines">
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_load" proved="true">
+ <goal name="vc_wrap_load_AtomicBool" proved="true">
+ <proof prover="0"><result status="valid" time="0.012441" steps="39494"/></proof>
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_load__refines">
+ <goal name="refines">
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_store">
+ <goal name="vc_wrap_store_AtomicBool">
+ </goal>
+</theory>
+<theory name="M_impl_AtomicExt_for_AtomicBool__wrap_store__refines">
+ <goal name="refines">
+ </goal>
+</theory>
+</file>
+</why3session>
